Fix the flexbox style for the header
Fixes the flexbox style for the header, margin and
padding for the unordered list in the header need
to be removed, and to ensure separtion of unordered
list links from the header from the rest of the page,
padding of 8 px has to be added. Then, 3 components
of unordered list left, unorfered list right and of
logo need to be styled as flexbox content with
align-items: center and justify-content:space between.
I couldn't guess almost any of it without looking at
the solution and got angry, probalby due to not having
much experience with flexbox and with the box model.
Wow! Great exercise.